Move `ClassDefinition` to the `syntax` package
